Ability to specify in Scan object max size of cache in bytes.
This allows to more finer-control over buffers and better separate load
when processing time is proportional to rows size, not their count.

For example, if we have lots of small rows which processed very fast and several
rows with lots of data which caused slower processing, buffer size allows to
better load separation over map jobs.