Fix HFile index overflow.
When index exceeds int (2^31) size, it is failed with error: https://issues.apache.org/jira/browse/HBASE-5071

This limitation was introduced by this optimisation: https://issues.apache.org/jira/browse/HBASE-3040

But we don't need to read whole index in memory, to decrease amount of reads. This fix uses BufferedInputStream to achieve this.